feat: Add purchase order due dates
Introduces the ability to set a due date when creating a new purchase order. This enhances tracking and management of outstanding supplier orders. The `PurchaseOrder` type and the `addPurchaseOrder` function have been updated to include an optional `dueDate` field. A default of 7 days is applied when a new order is created.