feat: Update app theme and title
Changes the default theme to light mode and updates the application title to "iCAN | F&B Costing Platform". Also updates the favicon color.

Introduces new types for StaffMember and Overhead and refactors the DataProvider to include mock data for these new types. This sets the foundation for managing staff and overhead costs within the platform.

Additionally, the `Recipe` type is enhanced with fields for labor, packaging, and wastage to support more detailed cost calculations.